+++ src/Telegram.Bots/Requests/IUserTargetable.cs
@@ -5,6 +5,6 @@ namespace Telegram.Bots.Requests
 {
   public interface IUserTargetable
   {
-    int UserId { get; }
+    long UserId { get; }
   }
 }
